requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core==1.12.0
azure-cosmos==4.2.0
azure-functions==1.7.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
six==1.15.0
urllib3==1.26.4
autopep8==1.5.6
azure-core==1.12.0
azure-cosmos==4.2.0
azure-functions==1.7.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.7.0
python-dotenv==0.16.0
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
